requirements.txt
# Основные зависимости
pyTelegramBotAPI
python-dotenv

# База данных
SQLAlchemy

# Админ-панель
sqladmin
starlette
uvicorn

# HTTP запросы для API
requests

# Дополнительные утилиты
python-dateutil
